Day 11 | Extracted relief strategy and switched back to Longs

+++ implementation/src/main/kotlin/io/github/tomplum/aoc/simluator/monkey/Monkey.kt
@@ -1,10 +1,8 @@
 package io.github.tomplum.aoc.simluator.monkey
 
-import java.math.BigInteger
-
 data class Monkey(
     val id: Int,
-    val items: MutableList<BigInteger>,
+    val items: MutableList<Long>,
     val operation: MonkeyOperation,
     val test: MonkeyTest
 )
