+++ implementation/src/main/kotlin/io/github/tomplum/aoc/simluator/monkey/MonkeyOperation.kt
@@ -1,15 +1,13 @@
 package io.github.tomplum.aoc.simluator.monkey
 
-import java.math.BigInteger
-
 class MonkeyOperation(instruction: String) {
     private val contents = instruction.trim().removePrefix("Operation: ").trim()
     private val end = contents.removePrefix("new = old ")
     private val operator = end.first()
     private val valueString = end.drop(1).trim()
 
-    fun execute(worryLevel: BigInteger): BigInteger {
-        val value = if (valueString == "old") worryLevel else valueString.toBigInteger()
+    fun execute(worryLevel: Long): Long {
+        val value = if (valueString == "old") worryLevel else valueString.toLong()
         return when(operator) {
             '+' -> worryLevel + value
             '*' -> worryLevel * value
